base de datos, clientes y propietarios quedan registrados.
