+++ src/main/java/openHouse/demo/entities/Owner.java
@@ -11,6 +11,7 @@ import java.util.List;
 public class Owner extends User{
 
     private String cbu;
+    private Boolean alta;
     
     @OneToMany
     private List<Property> propiedades;
@@ -33,6 +34,14 @@ public class Owner extends User{
 
     public void setPropiedades(List<Property> propiedades) {
         this.propiedades = propiedades;
+    }
+
+    public Boolean getAlta() {
+        return alta;
+    }
+
+    public void setAlta(Boolean alta) {
+        this.alta = alta;
     }
 
     @Override
