+++ src/main/resources/templates/fragments/errores.html
@@ -1,0 +1,12 @@
+<div th:if="${error != null}" class="card text-white bg-danger mb-3 mensajeError">
+    <div class="card-body">
+        <h5 class="card-title">Error!</h5>
+        <p class="card-text" th:text="${error}"></p>
+    </div>
+</div>
+<div th:if="${exito != null}" class="card text-white bg-success mb-3 mensajeError">
+    <div class="card-body">
+        <h5 class="card-title">Exito!</h5>
+        <p class="card-text" th:text="${exito}"></p>
+    </div>
+</div>
