+++ src/main/resources/templates/ingresar.html
@@ -19,8 +19,8 @@
                 <div class="message">
                     <h2>Bienvenido a Open-House</h2>
                     <p>Si ya tienes una cuenta por favor inicia sesion aqui</p>
-                    <a th:href="@{/registrarCliente}"><button class="sign-up-btn">Alquilar</button></a>
-                    <a th:href="@{/registrarPropietario}"><button class="sign-up-btn">Propietario</button></a>
+                    <a th:href="@{/cliente/registrarCliente}"><button class="sign-up-btn">Alquilar</button></a>
+                    <a th:href="@{/propietario/registrarPropietario}"><button class="sign-up-btn">Propietario</button></a>
                 </div>
             </div>
